Convert batch.x to bfloat16 instead of float

+++ graphgps/encoder/linear_node_encoder.py
@@ -11,6 +11,6 @@ class LinearNodeEncoder(torch.nn.Module):
         self.encoder = torch.nn.Linear(cfg.share.dim_in, emb_dim)
 
     def forward(self, batch):
-        batch.x = batch.x.float()
+        batch.x = batch.x.to(torch.bfloat16)
         batch.x = self.encoder(batch.x)
         return batch
